// packages/discord/src/operations/threadSearch.ts
export const ThreadSearchInput = /*@__PURE__*/ /*#__PURE__*/ Schema.Struct({
  channel_id: Schema.String.pipe(T.PathParam()),
  name: Schema.optional(Schema.String),
  slop: Schema.optional(Schema.Number),
  min_id: Schema.optional(Schema.String),
  max_id: Schema.optional(Schema.String),
  tag: Schema.optional(Schema.String),
  tag_setting: Schema.optional(Schema.String),
  archived: Schema.optional(Schema.Boolean),
  sort_by: Schema.optional(Schema.String),
  sort_order: Schema.optional(Schema.String),
  limit: Schema.optional(Schema.Number),
  offset: Schema.optional(Schema.Number),
}).pipe(
  T.Http({ method: "GET", path: "/channels/{channel_id}/threads/search" }),
);
export type ThreadSearchInput = typeof ThreadSearchInput.Type;
// ...
export type ThreadSearchOutput = typeof ThreadSearchOutput.Type;

// The operation
export const threadSearch = /*@__PURE__*/ /*#__PURE__*/ API.make(() => ({
  inputSchema: ThreadSearchInput,
  outputSchema: ThreadSearchOutput,
  errors: [BadRequest, Forbidden, NotFound] as const,
}));